update nampilin cover di card

+++ backend/src/catatanBelajar/index.ts
@@ -12,7 +12,9 @@ router.get('/catatanBelajars', async (req, res) => {
         include: {
             catatanbelajar_tag: {
                     include: {
-                        tag: true, // Include tag information
+                        tag: {
+                            select: { nama_tag: true } // Hanya memilih nama_tag
+                        }
                     }
                 }
         },
@@ -145,50 +147,53 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
     const catatanId = req.params.id;
     const updateCatatanData = req.body;
-    const tagNamas = updateCatatanData.nama_tag; // Ubah nama variabel agar sesuai dengan tipe data array
 
-    // Inisialisasi array untuk menyimpan ID tag baru
-    const newTagIds = [];
+    if (updateCatatanData.nama_tag != undefined) {
+        const tagNamas = updateCatatanData.nama_tag; // Ubah nama variabel agar sesuai dengan tipe data array
 
-    // Proses setiap nama tag
-    for (const tagNama of tagNamas) {
-        let tagId;
-        const existingTag = await prisma.tag.findUnique({
-            where: {
-                nama_tag: tagNama
-            }
-        });
+        // Inisialisasi array untuk menyimpan ID tag baru
+        const newTagIds = [];
 
-        if (existingTag) {
-            tagId = existingTag.id;
-        } else {
-            const newTag = await prisma.tag.create({
-                data: {
+        // Proses setiap nama tag
+        for (const tagNama of tagNamas) {
+            let tagId;
+            const existingTag = await prisma.tag.findUnique({
+                where: {
                     nama_tag: tagNama
                 }
             });
-            tagId = newTag.id;
+
+            if (existingTag) {
+                tagId = existingTag.id;
+            } else {
+                const newTag = await prisma.tag.create({
+                    data: {
+                        nama_tag: tagNama
+                    }
+                });
+                tagId = newTag.id;
+            }
+
+            // Tambahkan ID tag baru ke dalam array
+            newTagIds.push(tagId);
         }
 
-        // Tambahkan ID tag baru ke dalam array
-        newTagIds.push(tagId);
-    }
-
-    // Hapus semua tag lama dari catatan belajar
-    await prisma.catatanbelajar_tag.deleteMany({
-        where: {
-            id_catatanbelajar: parseInt(catatanId),
-        },
-    });
-
-    // Tambahkan tag baru ke catatan belajar
-    for (const tagId of newTagIds) {
-        await prisma.catatanbelajar_tag.create({
-            data: {
+        // Hapus semua tag lama dari catatan belajar
+        await prisma.catatanbelajar_tag.deleteMany({
+            where: {
                 id_catatanbelajar: parseInt(catatanId),
-                id_tag: tagId
-            }
+            },
         });
+
+        // Tambahkan tag baru ke catatan belajar
+        for (const tagId of newTagIds) {
+            await prisma.catatanbelajar_tag.create({
+                data: {
+                    id_catatanbelajar: parseInt(catatanId),
+                    id_tag: tagId
+                }
+            });
+        }
     }
 
     const updatedCatatanBelajar = await prisma.catatanbelajar.update({
